+++ CHANGELOG.md
@@ -64,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### shells.py
 - Fixes to boolean operations. Added some tolerance parameters to some methods. 
+- Shell3D: is_shell_open
 
 #### surfaces.py 
 - SphericalSurface3D: use circle 3d instead of polygon3D for plotting.